pyproject.toml
[tool.poetry]
name = "enviroplus-ql"
version = "0.4.0"
description = "A GraphQL server for querying Enviroplus sensor data"
authors = ["Ed Rutherford <erutherford@nullsecurity.tech>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql = "^0.247.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
asyncio = "^3.4.3"

[tool.poetry.scripts]
enviroplus-ql = "server.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
